Refactor: Move SharedState and ParkManager into their own files.
