+++ crates/lsio_threadpool/src/lib.rs
@@ -1,3 +1,4 @@
+pub(crate) mod park_manager;
+mod shared_state;
 pub mod threadpool;
-pub mod worker;
-
+pub(crate) mod worker;
