+++ crates/lsio_threadpool/src/threadpool.rs
@@ -1,117 +1,19 @@
 use std::{
-    collections::VecDeque,
     sync::{
         atomic::{AtomicBool, Ordering::Relaxed},
-        mpsc::{self, RecvError},
+        mpsc::{self},
         Arc,
     },
-    thread::{self, JoinHandle, Thread},
+    thread::{self, JoinHandle},
 };
 
 use crossbeam::deque;
 
-use crate::worker::WorkerThread;
-
-pub(crate) enum ParkManagerCommand {
-    WakeAtMostNThreads(u32),
-    ThreadIsParked(Thread),
-    Stop,
-}
-
-struct ParkManager {
-    rx: mpsc::Receiver<ParkManagerCommand>,
-    at_least_one_thread_is_parked: Arc<AtomicBool>,
-    parked_threads: VecDeque<Thread>,
-}
-
-impl ParkManager {
-    pub(crate) fn start(
-        rx: mpsc::Receiver<ParkManagerCommand>,
-        at_least_one_thread_is_parked: Arc<AtomicBool>,
-        n_worker_threads: usize,
-    ) -> JoinHandle<()> {
-        let mut park_manager = Self {
-            rx,
-            at_least_one_thread_is_parked,
-            parked_threads: VecDeque::with_capacity(n_worker_threads),
-        };
-        thread::Builder::new()
-            .name("ParkManager".to_string())
-            .spawn(move || park_manager.main_loop())
-            .expect("Failed to spawn the ParkManager thread!")
-    }
-
-    fn main_loop(&mut self) {
-        use ParkManagerCommand::*;
-        loop {
-            match self.rx.recv() {
-                Ok(cmd) => match cmd {
-                    ThreadIsParked(t) => self.thread_is_parked(t),
-                    WakeAtMostNThreads(n) => self.wake_at_most_n_threads(n),
-                    Stop => break,
-                },
-                Err(RecvError) => break,
-            }
-        }
-    }
-
-    fn thread_is_parked(&mut self, t: Thread) {
-        self.at_least_one_thread_is_parked.store(true, Relaxed);
-        debug_assert!(!self.parked_threads.iter().any(|pt| pt.id() == t.id()));
-        self.parked_threads.push_back(t);
-    }
-
-    fn wake_at_most_n_threads(&mut self, n: u32) {
-        for _ in 0..n {
-            match self.parked_threads.pop_front() {
-                Some(thread) => thread.unpark(),
-                None => break,
-            }
-        }
-        if self.parked_threads.is_empty() {
-            self.at_least_one_thread_is_parked.store(false, Relaxed);
-        }
-    }
-}
-
-/// Shared with worker threads
-#[derive(Debug)]
-pub(crate) struct Shared<T>
-where
-    T: Send,
-{
-    pub(crate) injector: Arc<deque::Injector<T>>,
-    pub(crate) keep_running: Arc<AtomicBool>,
-    pub(crate) chan_to_park_manager: mpsc::Sender<ParkManagerCommand>,
-    pub(crate) at_least_one_thread_is_parked: Arc<AtomicBool>,
-}
-
-impl<T> Shared<T>
-where
-    T: Send,
-{
-    pub(crate) fn unpark_at_most_n_threads(&self, n: u32) {
-        if self.at_least_one_thread_is_parked.load(Relaxed) {
-            self.chan_to_park_manager
-                .send(ParkManagerCommand::WakeAtMostNThreads(n))
-                .unwrap();
-        }
-    }
-}
-
-impl<T> Clone for Shared<T>
-where
-    T: Send,
-{
-    fn clone(&self) -> Self {
-        Self {
-            injector: Arc::clone(&self.injector),
-            keep_running: Arc::clone(&self.keep_running),
-            chan_to_park_manager: self.chan_to_park_manager.clone(),
-            at_least_one_thread_is_parked: Arc::clone(&self.at_least_one_thread_is_parked),
-        }
-    }
-}
+use crate::{
+    park_manager::{ParkManager, ParkManagerCommand},
+    shared_state::SharedState,
+    worker::WorkerThread,
+};
 
 #[derive(Debug)]
 pub struct ThreadPool<T>
@@ -120,7 +22,7 @@ where
 {
     /// thread_handles includes the worker threads and the ParkManager thread
     thread_handles: Vec<JoinHandle<()>>,
-    shared: Shared<T>,
+    shared: SharedState<T>,
 }
 
 impl<T> ThreadPool<T>
@@ -138,7 +40,7 @@ where
         OP: Fn(WorkerThread<T>) + Send + Clone + 'static,
     {
         let (chan_to_park_manager, rx_for_park_manager) = mpsc::channel();
-        let shared = Shared {
+        let shared = SharedState {
             injector: Arc::new(deque::Injector::new()),
             keep_running: Arc::new(AtomicBool::new(true)),
             chan_to_park_manager,
